fix errors when mutliple dhcp pids are running
as reported by "anonymous_user" on the arch bbs, if multiple
dhcpdc/dhclient processes are running, we could throw non-fatal errors
when checking for dhcp on the interface. this commit handles multiple
pids and checks them all until we find what we want. if we dont find
what we want, then we assume no dhcp on this interface.